Win32 compat fix Seems like my Cygwin+TortoiseGit setup messed up crlf data and now every while is changed. I did so many changes that I'm not going to manually fix this issue, and there doesn't seem to be automatic solution. Fuck you windows.

+++ alt/CoreAlt/Vector.h
@@ -726,7 +726,7 @@ public:
 		hash_t* it = begin;
 		while (it != end) {
 			if (*it == hash)
-				return it - begin;
+				return (int)(it - begin);
 			it++;
 		}
 		return -1;
